% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topslam.R
\name{topslam}
\alias{topslam}
\title{Run topslam}
\usage{
topslam(counts, start_cell_id, n_components = 2, n_neighbors = 10,
  linear_dims = 0, max_iters = 200, dimreds = c("t-SNE", "PCA",
  "Spectral", "Isomap", "ICA"), num_cores = 1)
}
\arguments{
\item{counts}{the counts}

\item{start_cell_id}{the names of starting cells}

\item{n_components}{the number of components}

\item{n_neighbors}{the number of neighbors}

\item{linear_dims}{?}

\item{max_iters}{the number of iterations to optimize over}

\item{dimreds}{the dimensionality reductions to use}

\item{num_cores}{the number of cores}
}
\description{
Run topslam
}
